feat: work on Contact CRUD

+++ app/Http/Controllers/ContactsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Contact;
+use http\Env\Response;
 
 class ContactsController extends Controller
 {
@@ -13,21 +14,30 @@ class ContactsController extends Controller
 
     public function store()
     {
-        Contact::create($this->validateData());
+        request()->user()->contacts()->create($this->validateData());
     }
 
     public function show(Contact $contact)
     {
+        if (request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
         return $contact;
     }
 
     public function update(Contact $contact)
     {
+        if (request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact)
     {
+        if (request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
         $contact->delete();
     }
